+++ src/ncea-enricher.tests/Services/XPathValidationServiceTests.cs
@@ -1,7 +1,5 @@
 ﻿using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
-using Microsoft.Extensions.Logging;
-using Moq;
 using Ncea.Enricher.Services;
 using Ncea.Enricher.Services.Contracts;
 using Ncea.Enricher.Tests.Clients;
@@ -12,14 +10,12 @@ namespace Ncea.Enricher.Tests.Services;
 public class XPathValidationServiceTests
 {
     private readonly IXmlValidationService _xpathValidationService;
-    private readonly Mock<ILogger<XPathValidationService>> _loggerMock;
     
     public XPathValidationServiceTests()
     {
-        LoggerForTests.Get(out _loggerMock);
         var serviceProvider = ServiceProviderForTests.Get();
         var configuration = serviceProvider!.GetService<IConfiguration>()!;
-        _xpathValidationService = new XPathValidationService(configuration, _loggerMock.Object);
+        _xpathValidationService = new XPathValidationService(configuration);
     }
 
     [Fact]
@@ -33,10 +29,5 @@ public class XPathValidationServiceTests
         _xpathValidationService.Validate(xDoc!);
 
         //Assert
-        _loggerMock.Verify(x => x.Log(LogLevel.Error,
-            It.IsAny<EventId>(),
-            It.IsAny<It.IsAnyType>(),
-            It.IsAny<Exception>(),
-            It.IsAny<Func<It.IsAnyType, Exception?, string>>()), Times.Exactly(0));
     }
 }
