--- src/ncea-enricher/Constants/ClassifierTypes.cs
@@ -1,9 +1,0 @@
-﻿namespace Ncea.Enricher.Constants
-{
-    public enum ClassifierTypes
-    {
-        Theme = 1,
-        Category = 2,
-        Subcategory = 3,
-    }
-}
