--- src/ncea-enricher/Constants/FieldType.cs
@@ -1,9 +1,0 @@
-﻿namespace Ncea.Enricher.Constants
-{
-    public enum FieldType
-    {
-        Text = 1,
-        List = 2,
-        ConditionalText = 3,
-    }
-}
