--- src/ncea-enricher/Constants/ResourceType.cs
@@ -1,9 +1,0 @@
-﻿namespace Ncea.Enricher.Constants
-{
-    public enum ResourceType
-    {
-        Dataset = 1,
-        Series = 2,
-        Service = 3,
-    }
-}
