+++ src/ncea-enricher/Enums/ClassifierTypes.cs
@@ -1,0 +1,8 @@
+﻿namespace Ncea.Enricher.Enums;
+
+public enum ClassifierTypes
+{
+    Theme = 1,
+    Category = 2,
+    Subcategory = 3,
+}
