+++ src/ncea-enricher/Enums/FieldName.cs
@@ -1,0 +1,30 @@
+﻿namespace Ncea.Enricher.Enums;
+
+public enum FieldName
+{
+    ResourceIdentifier = 1,
+    FileIdentifier = 2,
+    ResourceType = 3,
+    HierarchyLevel = 4,
+    Title = 5,
+    Abstract = 6,
+    Lineage = 7,
+    MetadataStandardName = 8,
+    MetadataStandardVersion = 9,
+    MetadataLanguage = 10,
+    MetadataDate = 11,
+    MetadataDateTime = 12,
+    MetadataDateTimeOffset = 13,
+    LimitationOnPublicAccess = 14,
+    License = 15,
+    ResponsiblePartyContact = 16,
+    ResponsiblePartyPointOfContact = 17,
+    FrequencyOfUpdate = 18,
+    AvailableFormats = 19,
+    AlternativeTitle = 20,
+    TopicCategory = 21,
+    Keywords = 22,
+    AdditionalInformation = 23,
+    GeographicExtentTitle = 24,
+    GeographicExtentCode = 25,
+}
