+++ src/ncea-enricher/Enums/FieldType.cs
@@ -1,0 +1,8 @@
+﻿namespace Ncea.Enricher.Enums;
+
+public enum FieldType
+{
+    Text = 1,
+    List = 2,
+    ConditionalText = 3,
+}
