+++ src/ncea-enricher/Enums/ResourceType.cs
@@ -1,0 +1,8 @@
+﻿namespace Ncea.Enricher.Enums;
+
+public enum ResourceType
+{
+    Dataset = 1,
+    Series = 2,
+    Service = 3,
+}
