--- src/ncea-enricher/Schema/gml/gml.xsd
@@ -1,20 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<schema targetNamespace="http://www.opengis.net/gml/3.2" xmlns:gml="http://www.opengis.net/gml/3.2" xmlns="http://www.w3.org/2001/XMLSchema" elementFormDefault="qualified" version="3.2.1.2">
-	<annotation>
-		<appinfo source="urn:x-ogc:specification:gml:schema-xsd:gml:3.2.1">gml.xsd</appinfo>
-		<documentation>
-			GML is an OGC Standard.
-			Copyright (c) 2007,2010 Open Geospatial Consortium.
-			To obtain additional rights of use, visit http://www.opengeospatial.org/legal/ .
-		</documentation>
-	</annotation>
-	<!-- ====================================================================== -->
-	<include schemaLocation="dynamicFeature.xsd"/>
-	<include schemaLocation="topology.xsd"/>
-	<include schemaLocation="coverage.xsd"/>
-	<include schemaLocation="coordinateReferenceSystems.xsd"/>
-	<include schemaLocation="observation.xsd"/>
-	<include schemaLocation="temporalReferenceSystems.xsd"/>
-	<include schemaLocation="deprecatedTypes.xsd"/>
-	<!-- ====================================================================== -->
-</schema>
